Removed already converted tests and fleshed LoadIsoautocorrCommandTests.
Files checked in: DataRepo/data/tests/isoautocorr/test-isoautocorr-study/test-isoautocorr-studydoc.xlsx DataRepo/tests/management/commands/legacy/test_load_accucor_msruns.py DataRepo/tests/management/commands/test_load_peak_annotations.py DataRepo/data/tests/isoautocorr/test-isoautocorr-study/test-isoautocorr-studydoc_v3.xlsx
