+++ DataRepo/tests/management/commands/legacy/test_load_accucor_msruns.py
@@ -2,9 +2,7 @@ import hashlib
 from pathlib import Path
 
 import pandas as pd
-from django.conf import settings
-from django.core.management import CommandError, call_command
-from django.test import override_settings, tag
+from django.core.management import call_command
 
 from DataRepo.loaders.legacy.accucor_data_loader import (
     AccuCorDataLoader,
@@ -15,554 +13,16 @@ from DataRepo.models import (
     ArchiveFile,
     DataFormat,
     Infusate,
-    InfusateTracer,
     MaintainedModel,
     MSRunSample,
     MSRunSequence,
-    PeakData,
     PeakGroup,
-    Sample,
-    Tracer,
-    TracerLabel,
 )
 from DataRepo.models.study import Study
 from DataRepo.models.utilities import exists_in_db
 from DataRepo.tests.tracebase_test_case import TracebaseTestCase
 from DataRepo.utils import read_from_file
 from DataRepo.utils.infusate_name_parser import parse_infusate_name_with_concs
-
-
-@override_settings(CACHES=settings.TEST_CACHES)
-class IsoCorrDataLoadingTests(TracebaseTestCase):
-    fixtures = ["data_types.yaml", "data_formats.yaml", "lc_methods.yaml"]
-
-    @classmethod
-    @MaintainedModel.no_autoupdates()
-    def setUpTestData(cls):
-        call_command(
-            "legacy_load_study",
-            "DataRepo/data/tests/protocols/loading.yaml",
-            verbosity=2,
-        )
-        call_command(
-            "legacy_load_study",
-            "DataRepo/data/tests/tissues/loading.yaml",
-            verbosity=2,
-        )
-        call_command(
-            "load_compounds",
-            infile="DataRepo/data/tests/compounds/consolidated_tracebase_compound_list.tsv",
-            verbosity=2,
-        )
-
-        call_command(
-            "legacy_load_animals_and_samples",
-            animal_and_sample_table_filename=(
-                "DataRepo/data/tests/singly_labeled_isocorr/animals_samples.xlsx"
-            ),
-            skip_researcher_check=True,
-        )
-
-        super().setUpTestData()
-
-    @MaintainedModel.no_autoupdates()
-    def load_multitracer_data(self):
-        call_command(
-            "legacy_load_animals_and_samples",
-            animal_and_sample_table_filename="DataRepo/data/tests/multiple_tracers/animal_sample_table.xlsx",
-            skip_researcher_check=True,
-        )
-
-        num_samples = 4
-        num_infusates = 2
-        num_infusatetracers = 3
-        num_tracers = 6
-        num_tracerlabels = 12
-
-        return (
-            num_samples,
-            num_infusates,
-            num_infusatetracers,
-            num_tracers,
-            num_tracerlabels,
-        )
-
-    @MaintainedModel.no_autoupdates()
-    def load_multilabel_data(self):
-        call_command(
-            "legacy_load_animals_and_samples",
-            animal_and_sample_table_filename=(
-                "DataRepo/data/tests/multiple_labels/animal_sample_table.xlsx"
-            ),
-            skip_researcher_check=True,
-        )
-
-        num_samples = 5
-        num_infusates = 2
-        num_infusatetracers = 2
-        num_tracers = 2
-        num_tracerlabels = 3
-
-        return (
-            num_samples,
-            num_infusates,
-            num_infusatetracers,
-            num_tracers,
-            num_tracerlabels,
-        )
-
-    @MaintainedModel.no_autoupdates()
-    def test_singly_labeled_isocorr_load(self):
-        pre_pg_load_count = PeakGroup.objects.count()
-        call_command(
-            "legacy_load_accucor_msruns",
-            accucor_file="DataRepo/data/tests/singly_labeled_isocorr/small_cor.csv",
-            skip_samples=("Blank01",),
-            lc_protocol_name="polar-HILIC-25-min",
-            instrument="unknown",
-            date="2021-04-29",
-            researcher="Michael Neinast",
-            new_researcher=True,
-            isocorr_format=True,
-        )
-        post_pg_load_count = PeakGroup.objects.count()
-        # The number of samples in the isocorr csv file (not the samples file)
-        SAMPLES_COUNT = 3
-        PEAKDATA_ROWS = 6
-        MEASURED_COMPOUNDS_COUNT = 2
-
-        self.assertEqual(
-            post_pg_load_count - pre_pg_load_count,
-            MEASURED_COMPOUNDS_COUNT * SAMPLES_COUNT,
-            msg=f"PeakGroup record count should be the number of compounds [{MEASURED_COMPOUNDS_COUNT}] times the "
-            f"number of samples [{SAMPLES_COUNT}] = [{MEASURED_COMPOUNDS_COUNT * SAMPLES_COUNT}].",
-        )
-        self.assertEqual(
-            PeakData.objects.all().count(),
-            PEAKDATA_ROWS * SAMPLES_COUNT,
-            msg=f"PeakData record count should be the number of peakdata rows [{PEAKDATA_ROWS}] times the number of "
-            f"samples [{SAMPLES_COUNT}] = [{PEAKDATA_ROWS * SAMPLES_COUNT}].",
-        )
-
-    @MaintainedModel.no_autoupdates()
-    def test_singly_labeled_isocorr_missing_flag_error(self):
-        """
-        Test to make sure the data-format option is suggested when not supplied
-        """
-        with self.assertRaises(CommandError) as ce:
-            call_command(
-                "legacy_load_accucor_msruns",
-                accucor_file="DataRepo/data/tests/singly_labeled_isocorr/small_cor.csv",
-                skip_samples=("Blank01",),
-                lc_protocol_name="polar-HILIC-25-min",
-                instrument="unknown",
-                date="2021-04-29",
-                researcher="Michael Neinast",
-                new_researcher=True,
-            )
-        self.assertIn(
-            "--data-format",
-            repr(ce.exception),
-            msg=f"This error should reference --data-format: [{ce.exception}]",
-        )
-
-    def get_model_counts(self):
-        return (
-            Sample.objects.count(),
-            Infusate.objects.count(),
-            InfusateTracer.objects.count(),
-            Tracer.objects.count(),
-            TracerLabel.objects.count(),
-        )
-
-    @MaintainedModel.no_autoupdates()
-    def test_multitracer_sample_table_load(self):
-        num_samples = 4
-        num_infusates = 2
-        num_infusatetracers = 9
-        num_tracers = 9
-        num_tracerlabels = 12
-
-        (
-            pre_samples,
-            pre_infusates,
-            pre_inftrcs,
-            pre_tracers,
-            pre_trclbls,
-        ) = self.get_model_counts()
-
-        call_command(
-            "legacy_load_animals_and_samples",
-            animal_and_sample_table_filename="DataRepo/data/tests/multiple_tracers/animal_sample_table.xlsx",
-            skip_researcher_check=True,
-        )
-
-        (
-            post_samples,
-            post_infusates,
-            post_inftrcs,
-            post_tracers,
-            post_trclbls,
-        ) = self.get_model_counts()
-
-        self.assert_model_counts(
-            num_samples,
-            num_infusates,
-            num_infusatetracers,
-            num_tracers,
-            num_tracerlabels,
-            pre_samples,
-            pre_infusates,
-            pre_inftrcs,
-            pre_tracers,
-            pre_trclbls,
-            post_samples,
-            post_infusates,
-            post_inftrcs,
-            post_tracers,
-            post_trclbls,
-        )
-
-    def assert_model_counts(
-        self,
-        num_samples,
-        num_infusates,
-        num_infusatetracers,
-        num_tracers,
-        num_tracerlabels,
-        pre_samples,
-        pre_infusates,
-        pre_inftrcs,
-        pre_tracers,
-        pre_trclbls,
-        post_samples,
-        post_infusates,
-        post_inftrcs,
-        post_tracers,
-        post_trclbls,
-    ):
-        self.assertEqual(num_samples, post_samples - pre_samples)
-        self.assertEqual(num_infusates, post_infusates - pre_infusates)
-        self.assertEqual(
-            num_infusatetracers,
-            post_inftrcs - pre_inftrcs,
-        )
-        self.assertEqual(num_tracers, post_tracers - pre_tracers)
-        self.assertEqual(num_tracerlabels, post_trclbls - pre_trclbls)
-
-    def assert_group_data_sample_counts(
-        self,
-        SAMPLES_COUNT,
-        PEAKDATA_ROWS,
-        PARENT_REC_COUNT,
-        pre_load_group_count,
-        post_load_group_count,
-    ):
-        self.assertEqual(
-            post_load_group_count - pre_load_group_count,
-            PARENT_REC_COUNT * SAMPLES_COUNT,
-            msg=f"PeakGroup record count should be the number of C12 PARENT lines [{PARENT_REC_COUNT}] times the "
-            f"number of samples [{SAMPLES_COUNT}] = [{PARENT_REC_COUNT * SAMPLES_COUNT}].",
-        )
-        self.assertEqual(
-            PeakData.objects.count(),
-            PEAKDATA_ROWS * SAMPLES_COUNT,
-            msg=f"PeakData record count should be the number of peakdata rows [{PEAKDATA_ROWS}] times the number of "
-            f"samples [{SAMPLES_COUNT}] = [{PEAKDATA_ROWS * SAMPLES_COUNT}].",
-        )
-
-    @MaintainedModel.no_autoupdates()
-    def test_multitracer_isocorr_load_1(self):
-        self.load_multitracer_data()
-        pre_load_group_count = PeakGroup.objects.count()
-        call_command(
-            "legacy_load_accucor_msruns",
-            accucor_file="DataRepo/data/tests/multiple_tracers/6eaafasted1_cor.xlsx",
-            lc_protocol_name="polar-HILIC-25-min",
-            instrument="unknown",
-            date="2021-04-29",
-            researcher="Xianfeng Zeng",
-            new_researcher=False,
-            isocorr_format=True,
-        )
-        post_load_group_count = PeakGroup.objects.count()
-        # The number of samples in the isocorr xlsx file (not the samples file)
-        SAMPLES_COUNT = 2
-        PEAKDATA_ROWS = 21
-        PARENT_REC_COUNT = 3
-        self.assert_group_data_sample_counts(
-            SAMPLES_COUNT,
-            PEAKDATA_ROWS,
-            PARENT_REC_COUNT,
-            pre_load_group_count,
-            post_load_group_count,
-        )
-
-    @MaintainedModel.no_autoupdates()
-    def test_multitracer_isocorr_load_2(self):
-        self.load_multitracer_data()
-        pre_load_group_count = PeakGroup.objects.count()
-        call_command(
-            "legacy_load_accucor_msruns",
-            accucor_file="DataRepo/data/tests/multiple_tracers/bcaafasted_cor.xlsx",
-            lc_protocol_name="polar-HILIC-25-min",
-            instrument="unknown",
-            date="2021-04-29",
-            researcher="Xianfeng Zeng",
-            new_researcher=False,
-            isocorr_format=True,
-        )
-        post_load_group_count = PeakGroup.objects.count()
-        # The number of samples in the isocorr xlsx file (not the samples file)
-        SAMPLES_COUNT = 2
-        PEAKDATA_ROWS = 24
-        PARENT_REC_COUNT = 2
-        self.assert_group_data_sample_counts(
-            SAMPLES_COUNT,
-            PEAKDATA_ROWS,
-            PARENT_REC_COUNT,
-            pre_load_group_count,
-            post_load_group_count,
-        )
-
-    @MaintainedModel.no_autoupdates()
-    def test_multilabel_sample_table_load(self):
-        num_samples = 6
-        num_infusates = 2
-        num_infusatetracers = 2
-        num_tracers = 2
-        num_tracerlabels = 4  # TracerLabel records are not unique. Note there would be 3 unique label records
-
-        (
-            pre_samples,
-            pre_infusates,
-            pre_inftrcs,
-            pre_tracers,
-            pre_trclbls,
-        ) = self.get_model_counts()
-
-        call_command(
-            "legacy_load_animals_and_samples",
-            animal_and_sample_table_filename=(
-                "DataRepo/data/tests/multiple_labels/animal_sample_table.xlsx"
-            ),
-            skip_researcher_check=True,
-        )
-
-        (
-            post_samples,
-            post_infusates,
-            post_inftrcs,
-            post_tracers,
-            post_trclbls,
-        ) = self.get_model_counts()
-
-        # Comment to break up JSCPD detection of clones
-
-        self.assert_model_counts(
-            num_samples,
-            num_infusates,
-            num_infusatetracers,
-            num_tracers,
-            num_tracerlabels,
-            pre_samples,
-            pre_infusates,
-            pre_inftrcs,
-            pre_tracers,
-            pre_trclbls,
-            post_samples,
-            post_infusates,
-            post_inftrcs,
-            post_tracers,
-            post_trclbls,
-        )
-
-    @MaintainedModel.no_autoupdates()
-    def test_multilabel_isocorr_load_1(self):
-        self.load_multilabel_data()
-        pre_load_group_count = PeakGroup.objects.count()
-        call_command(
-            "legacy_load_accucor_msruns",
-            accucor_file="DataRepo/data/tests/multiple_labels/alafasted_cor.xlsx",
-            lc_protocol_name="polar-HILIC-25-min",
-            instrument="unknown",
-            date="2021-04-29",
-            researcher="Xianfeng Zeng",
-            new_researcher=False,
-            isocorr_format=True,
-        )
-        post_load_group_count = PeakGroup.objects.count()
-
-        self.assert_peak_group_counts(
-            pre_load_group_count, post_load_group_count, 4, 37, 4
-        )
-
-    @MaintainedModel.no_autoupdates()
-    def test_multilabel_isocorr_load_2(self):
-        self.load_multilabel_data()
-        pre_load_group_count = PeakGroup.objects.count()
-        call_command(
-            "legacy_load_accucor_msruns",
-            accucor_file="DataRepo/data/tests/multiple_labels/glnfasted1_cor.xlsx",
-            lc_protocol_name="polar-HILIC-25-min",
-            instrument="unknown",
-            date="2021-04-29",
-            researcher="Xianfeng Zeng",
-            new_researcher=False,
-            isocorr_format=True,
-        )
-        post_load_group_count = PeakGroup.objects.count()
-
-        self.assert_peak_group_counts(
-            pre_load_group_count, post_load_group_count, 2, 26, 2
-        )
-
-    def assert_peak_group_counts(
-        self,
-        pre_load_group_count,
-        post_load_group_count,
-        samples_count,
-        peakdata_rows,
-        parent_rec_count,
-    ):
-        self.assertEqual(
-            post_load_group_count - pre_load_group_count,
-            parent_rec_count * samples_count,
-            msg=f"PeakGroup record count should be the number of C12 PARENT lines [{parent_rec_count}] times the "
-            f"number of samples [{samples_count}] = [{parent_rec_count * samples_count}].",
-        )
-        self.assertEqual(
-            PeakData.objects.count(),
-            peakdata_rows * samples_count,
-            msg=f"PeakData record count should be the number of peakdata rows [{peakdata_rows}] times the number of "
-            f"samples [{samples_count}] = [{peakdata_rows * samples_count}].",
-        )
-
-    @MaintainedModel.no_autoupdates()
-    def test_labeled_elements_common_with_compound(self):
-        """
-        Test to ensure count 0 entries are not created when measured compound doesn't have that element
-        """
-        self.load_multilabel_data()
-        call_command(
-            "legacy_load_accucor_msruns",
-            accucor_file="DataRepo/data/tests/multiple_labels/alafasted_cor.xlsx",
-            lc_protocol_name="polar-HILIC-25-min",
-            instrument="unknown",
-            date="2021-04-29",
-            researcher="Xianfeng Zeng",
-            new_researcher=False,
-            isocorr_format=True,
-        )
-        pg = (
-            PeakGroup.objects.filter(msrun_sample__sample__name="xzl5_panc")
-            .filter(name__exact="serine")
-            .filter(
-                peak_annotation_file__filename="alafasted_cor.xlsx",
-            )
-            .order_by("id", "peak_data__labels__element")
-            .distinct("id", "peak_data__labels__element")
-        )
-
-        self.assertEqual(pg.count(), 2)
-        self.assertEqual(pg.filter(peak_data__labels__element__exact="C").count(), 1)
-        self.assertEqual(pg.filter(peak_data__labels__element__exact="N").count(), 1)
-
-
-@tag("isoautocorr")
-@override_settings(CACHES=settings.TEST_CACHES)
-class IsoAutoCorrDataLoadingTests(TracebaseTestCase):
-    fixtures = ["data_types.yaml", "data_formats.yaml", "lc_methods.yaml"]
-
-    @classmethod
-    def setUpTestData(cls):
-        call_command(
-            "legacy_load_study",
-            "DataRepo/data/tests/protocols/loading.yaml",
-            verbosity=2,
-        )
-        call_command(
-            "legacy_load_study",
-            "DataRepo/data/tests/tissues/loading.yaml",
-            verbosity=2,
-        )
-        call_command(
-            "load_compounds",
-            infile="DataRepo/data/tests/compounds/consolidated_tracebase_compound_list.tsv",
-            verbosity=2,
-        )
-        call_command(
-            "legacy_load_animals_and_samples",
-            animal_and_sample_table_filename=(
-                "DataRepo/data/tests/isoautocorr/test-isoautocorr-study/test-isoautocorr-studydoc.xlsx"
-            ),
-        )
-        call_command(
-            "legacy_load_accucor_msruns",
-            accucor_file="DataRepo/data/tests/isoautocorr/test-isoautocorr-study/test-isoautocorr-negative-cor.xlsx",
-            lc_protocol_name="polar-HILIC-25-min",
-            instrument="Exploris480",
-            date="2024-05-23",
-            researcher="Michael Neinast",
-            new_researcher=False,
-        )
-        cls.SAMPLES_COUNT = 4
-        cls.PEAKDATA_ROWS = 14
-        cls.MEASURED_COMPOUNDS_COUNT = 2  # L-Serine and Glucose
-
-        super().setUpTestData()
-
-    def test_isoautocorr_load(self):
-        """Load test-isoautocorr data"""
-
-        self.assertEqual(
-            PeakGroup.objects.count(),
-            self.MEASURED_COMPOUNDS_COUNT * self.SAMPLES_COUNT,
-        )
-        self.assertEqual(
-            PeakData.objects.all().count(), self.PEAKDATA_ROWS * self.SAMPLES_COUNT
-        )
-
-    def test_isoautocorr_peakdatalabels(self):
-        """Check peak data labels for isoautocorr loading"""
-
-        peak_data = PeakData.objects.filter(peak_group__name="Glycine").filter(
-            peak_group__msrun_sample__sample__name="His_neg_M3_T02_liv"
-        )
-
-        peak_data_labels = []
-        for peakdata in peak_data.all():
-            pdl = peakdata.labels.values("element", "mass_number", "count")
-            peak_data_labels.append(list(pdl))
-
-        expected = [
-            [
-                {"element": "C", "mass_number": 13, "count": 0},
-                {"element": "N", "mass_number": 15, "count": 0},
-            ],
-            [
-                {"element": "C", "mass_number": 13, "count": 0},
-                {"element": "N", "mass_number": 15, "count": 1},
-            ],
-            [
-                {"element": "C", "mass_number": 13, "count": 2},
-                {"element": "N", "mass_number": 15, "count": 0},
-            ],
-            [
-                {"element": "C", "mass_number": 13, "count": 1},
-                {"element": "N", "mass_number": 15, "count": 0},
-            ],
-            [
-                {"element": "C", "mass_number": 13, "count": 1},
-                {"element": "N", "mass_number": 15, "count": 1},
-            ],
-            [
-                {"element": "C", "mass_number": 13, "count": 2},
-                {"element": "N", "mass_number": 15, "count": 1},
-            ],
-        ]
-
-        self.assertEqual(expected, list(peak_data_labels))
 
 
 class MSRunSampleSequenceTests(TracebaseTestCase):
